# omnigram.openapi.spec.yaml
openapi: 3.0.1
info:
  title: omnigram
  description: ''
  version: 1.0.0
...
paths:
...
  /reader/books:
    get:
      summary: 获取书籍列表
      deprecated: false
      description: ''
      tags: []
      parameters:
        - name: search
          in: query
          description: 模糊搜索字段
          required: false
          example: ''
          schema:
            type: string
        - name: page_size
          in: query
          description: 页数
          required: false
          schema:
            type: integer
        - name: page_num
          in: query
          description: 页数大小
          required: false
          schema:
            type: integer
        - name: category
          in: query
          description: 分类
          required: false
          schema:
            type: string
        - name: author
          in: query
          description: 作者
          required: false
          schema:
            type: string
      responses:
        '200':
          description: 获取书籍列表
          content:
            application/json:
              schema: &ref_4
                $ref: '#/components/schemas/ebook_list_dto'
              examples:
                '1':
                  summary: 成功示例
                  value:
                    total: 51
                    items:
                      - id: 79
                        size: 45
                        ctime: '2017-09-22 06:12:40'
                        utime: '1972-08-03 06:00:09'
                        title: 书何效内当
                        sub_title: 律给将离志转问
                        language: laboris labore dolor ipsum voluptate
                        cover_url: http://clsn.sr/ybahcn
                        uuid: '55'
                        isbn: adipisicing minim qui laboris
                        asin: eiusmod ea
                        identifier: '11'
                        category: tempor ut Excepteur consequat dolor
                        author: proident Excepteur dolore
                        author_url: http://lkvmhd.com.cn/rtjwlvd
                        author_sort: ex ut amet aute
                        publisher: officia consectetur exercitation amet deserunt
                        description: >-
                          记主导反分圆们照何办两通品容积。处特明际百比组儿想将油拉该。时作权月不段三导组他样加育。收究过达他无花求心所如科认团。根国想把家引建各情级装后图委管生名。
                        tags:
                          - Ut
                          - commodo deserunt Duis anim
                        pubdate: '1975-03-27'
                        rating: 20
                        publisher_url: http://txbjtyp.jm/ktebevpfd
                        count_visit: 50
                        count_download: 95
                        progress: 3
                        progress_index: 12
                        para_position: 53
          headers: {}
      security:
        - bearer: []
...
  /reader/books/{book_id}:
    get:
      summary: 书籍详情
      deprecated: false
      description: ''
      tags: []
      parameters:
        - name: book_id
          in: path
          description: ''
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 数据详情
          content:
            application/json:
              schema: &ref_5
                $ref: '#/components/schemas/ebook_dto'
              examples:
                '1':
                  summary: 成功示例
                  value:
                    code: 200
                    message: ok
                    data:
                      total: 100
                      items:
                        - id: 12345
                          size: 54321
                          ctime: '2024-08-15T12:00:00Z'
                          utime: '2024-08-15T12:00:00Z'
                          title: Golang Programming
                          sub_title: Mastering Go
                          language: English
                          cover_url: https://example.com/cover.jpg
                          uuid: 123e4567-e89b-12d3-a456-426614174000
                          isbn: 978-3-16-148410-0
                          asin: B01N5K7PCC
                          identifier: unique-identifier-001
                          category: Programming
                          author: John Doe
                          author_url: https://example.com/author
                          author_sort: Doe, John
                          publisher: Tech Books Publishing
                          description: A comprehensive guide to Go programming.
                          tags:
                            - go
                            - programming
                            - tech
                          pubdate: '2023-07-15'
                          rating: 4.5
                          publisher_url: https://example.com/publisher
                          count_visit: 1000
                          count_download: 500
                          progress: 0.1
                          progress_index: 100
                          para_position: 100
          headers: {}
      security:
        - bearer: []
    put:
      summary: 修改书籍介绍信息
      deprecated: false
      description: ''
      tags: []
      parameters:
        - name: book_id
          in: path
          description: ''
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  title: 标题
                sub_title:
                  type: string
                  title: 子标题
                language:
                  type: string
                  title: 图书语言
                cover_url:
                  type: string
                  title: 封面URL
                isbn:
                  type: string
                  title: ISBN
                asin:
                  type: string
                  title: AWS ID
                category:
                  type: string
                  title: 分类
                author:
                  type: string
                  title: 作者
                author_url:
                  type: string
                  title: 作者URL
                author_sort:
                  type: string
                  title: 作者列表
                publisher:
                  type: string
                  title: 出版社
                description:
                  type: string
                  title: 描述
                tags:
                  type: array
                  items:
                    type: string
                  title: 标签
                pubdate:
                  type: string
                  title: 出版日期
                rating:
                  type: number
                  title: 评分
                publisher_url:
                  type: string
                  title: 出版链接
              required:
                - title
                - sub_title
                - language
            example:
              id: 12345
              title: Golang Programming
              sub_title: Mastering Go
              language: English
              cover_url: https://example.com/cover.jpg
              isbn: 978-3-16-148410-0
              asin: B01N5K7PCC
              category: Programming
              author: John Doe
              author_url: https://example.com/author
              author_sort: Doe, John
              publisher: Tech Books Publishing
              description: A comprehensive guide to Go programming.
              tags:
                - go
                - programming
                - tech
              pubdate: '2023-07-15'
              rating: 4.5
              publisher_url: https://example.com/publisher
      responses:
        '200':
          description: 获取书面详情
          content:
            application/json:
              schema: *ref_5
              examples:
                '1':
                  summary: 成功示例
                  value:
                    code: 200
                    message: ok
                    data:
                      total: 100
                      items:
                        - id: 12345
                          size: 54321
                          ctime: '2024-08-15T12:00:00Z'
                          utime: '2024-08-15T12:00:00Z'
                          title: Golang Programming
                          sub_title: Mastering Go
                          language: English
                          cover_url: https://example.com/cover.jpg
                          uuid: 123e4567-e89b-12d3-a456-426614174000
                          isbn: 978-3-16-148410-0
                          asin: B01N5K7PCC
                          identifier: unique-identifier-001
                          category: Programming
                          author: John Doe
                          author_url: https://example.com/author
                          author_sort: Doe, John
                          publisher: Tech Books Publishing
                          description: A comprehensive guide to Go programming.
                          tags:
                            - go
                            - programming
                            - tech
                          pubdate: '2023-07-15'
                          rating: 4.5
                          publisher_url: https://example.com/publisher
                          count_visit: 1000
                          count_download: 500
                          progress: 0.1
                          progress_index: 100
                          para_position: 100
          headers: {}
      security:
        - bearer: []
...
  /reader/download/books/{book_id}:
    get:
      summary: 下载书籍
      deprecated: false
      description: ''
      tags: []
      parameters:
        - name: book_id
          in: path
          description: ''
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 下载书
          content:
            application/octet-stream:
              schema:
                type: object
                properties: {}
          headers: {}
      security:
        - bearer: []
...
components:
  schemas:
...
    ebook_dto:
      type: object
      properties:
        id:
          type: integer
          title: 书ID
        size:
          type: integer
          title: 文件大小
        ctime:
          type: string
          title: 创建时间
        utime:
          type: string
          title: 更新时间
        title:
          type: string
          title: 标题
        sub_title:
          type: string
          title: 子标题
        language:
          type: string
          title: 图书语言
        cover_url:
          type: string
          title: 封面URL
        uuid:
          type: string
          title: 图书UUID
        isbn:
          type: string
          title: ISBN
        asin:
          type: string
          title: AWS ID
        identifier:
          type: string
          title: 唯一ID
        category:
          type: string
          title: 分类
        author:
          type: string
          title: 作者
        author_url:
          type: string
          title: 作者URL
        author_sort:
          type: string
          title: 作者列表
        publisher:
          type: string
          title: 出版社
        description:
          type: string
          title: 描述
        fav_status:
          type: boolean
          title: 喜欢状态
        tags:
          type: array
          items:
            type: string
          title: 标签
        pubdate:
          type: string
          title: 出版日期
        rating:
          type: number
          title: 评分
        publisher_url:
          type: string
          title: 出版链接
        count_visit:
          type: integer
          title: 访问次数
        count_download:
          type: integer
          title: 下载次数
        progress:
          type: number
          title: 阅读进度
        progress_index:
          type: integer
          title: 阅读进度索引
        para_position:
          type: integer
          title: 段落位置
      required:
        - id
        - title
        - identifier
...
    ebook_list_dto:
      type: object
      properties:
        total:
          type: integer
        items:
          type: array
          items: *ref_5
      required:
        - total
        - items
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
servers: []